+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 * Add a `Stream.slices()` extension method.
 
+* Fix a bug where `CancelableOperation.then` may invoke the `onValue` callback,
+  even if it had been canceled before `CancelableOperation.value` completes.
+
 ## 2.6.1
 
 * When `StreamGroup.stream.listen()` is called, gracefully handle component